requirements.txt
anyio==4.3.0
blinker==1.4
certifi==2024.2.2
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-telegram-bot==21.1.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
systemd-python==234
typing_extensions==4.11.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
